+++ templates/question.html
@@ -9,77 +9,80 @@
 <table border="2" cellpadding="5" cellspacing="" width="1200" style="border-collapse:collapse;">
     <thead>
     <tr>
-        {% for item in q_fields %}
-        <th>{{ item }}</th>
+        {% for item in question_headers %}
+            <th>{{ item }}</th>
         {% endfor %}
     </tr>
     </thead>
     <tbody>
     <tr>
-        {% for question in questions %}
-        {% if question['id'] == quest_id %}
-    <tr>
-        {% for key, value in question.items() %}
-        {% if key != 'id' and key != 'image'%}
-        <td align="center">{{ value }}</td>
-        {% endif %}
-        {% endfor %}
-    </tr>
-
-    </tbody>
-</table>
-<br>
-<form action="/question/{{ question['id'] }}/new-answer"><span style="float:left;">
+        {% if questions %}
+            {% for question in questions %}
+                {% if quest_id == question.id %}
+                    <tr>
+                        {% for key, value in question.items() %}
+                            {% if key != 'id' and key != 'image' %}
+                                <td align="center">{{ value }}</td>
+                            {% endif %}
+                        {% endfor %}
+                    </tr>
+                    </tbody>
+                    </table>
+                    <br>
+                    <form action="/question/{{ question.id }}/new-answer"><span style="float:left;">
         <button type="submit"
                 style="margin: 10px; width:130px; height:25px"><strong>Add new answer</strong></button></span>
-</form>
+                    </form>
 
-<form action="/question/{{ question['id'] }}/edit"><span style="float:left;">
+                    <form action="/question/{{ question.id }}/edit"><span style="float:left;">
         <button type="submit"
                 style="margin:10px; width:130px; height:25px"><strong>Edit Question</strong></button></span>
-</form>
-<form action="/question/{{ question['id'] }}/delete" method="post"><span style="float:left;">
+                    </form>
+                    <form action="/question/{{ question.id }}/delete" method="post"><span style="float:left;">
         <button type="submit"
                 style="margin:10px; width:130px; height:25px"><strong>Delete Question</strong></button></span>
-</form>
-{% endif %}
-{% endfor %}
-<br>
-{% if any_answer %}
-<br>
-<h1><p>Answers:</p></h1>
-<table border="2" cellpadding="5" cellspacing="" width="1200" style="border-collapse:collapse;">
-    <thead>
-    <tr>
-        {% for item in a_fields %}
-        <th>{{ item }}</th>
-        {% endfor %}
-    </tr>
-    </thead>
-    <tbody>
-    <tr>
-        {% for answer in answers %}
-        {% if answer['question_id'] == quest_id %}
-    <tr>
-        {% for key, value in answer.items() %}
-        {% if key != 'id' and key != 'question_id' and key != 'image'%}
-        <td align="center">{{ value }}</td>
+                    </form>
+                {% endif %}
+            {% endfor %}
+            <br>
         {% endif %}
-        {% endfor %}
-        <td>
-            <form action="/answer/{{ answer['id'] }}/delete" align="center" method="post">
-                <button type="submit" style="width:80px; height:25px"><strong>Delete</strong></button>
-            </form>
-        </td>
-    </tr>
-    {% endif %}
-    {% endfor %}
-    </tbody>
-</table>
-{% else %}
 <br>
-<h1>Answers: </h1>
-<h2>There are no answers to this question yet</h2>
+{% if is_answer %}
+    <h1>Answers:</h1>
+    <table border="2" cellpadding="5" cellspacing="" width="1200" style="border-collapse:collapse;">
+        <thead>
+        <tr>
+            {% for item in answer_headers %}
+                <th>{{ item }}</th>
+            {% endfor %}
+        </tr>
+        </thead>
+        <tbody>
+        <tr>
+            {% if all_answer %}
+                {% for answer in all_answer %}
+                    {% if quest_id == answer.question_id %}
+                        <tr>
+                            {% for key, value in answer.items() %}
+                                {% if key != 'id' and key != 'image' and key != 'question_id' %}
+                                    <td align="center">{{ value }}</td>
+                                {% endif %}
+                            {% endfor %}
+                            <td>
+                                <form action="/answer/{{ answer.id }}/delete" align="center" method="post">
+                                    <button type="submit" style="width:80px; height:25px"><strong>Delete</strong>
+                                    </button>
+                                </form>
+                            </td>
+                        </tr>
+                    {% endif %}
+                {% endfor %}
+            {% endif %}
+        </tbody>
+    </table>
+{% else %}
+    <h1>Answers: </h1>
+    <h2>There are no answers to this questions yet.</h2>
 {% endif %}
 <br><br>
 <form action="/list" >
